Add functionality for writing documents to elasticsearch
This matter is covered in issue #4. In this commit, I added two lines to the write function. The first line creates a document in elasticsearch from specified keyword parameters. The second line tells elasticsearch to refresh the index we just wrote to it, so we can be sure the document we index will be immediately availible to search queries.