:bug: :x: :exclamation: [QASM] Critical Bugfix: Use Correct Rotation Gate Spec
This is a simple bugfix that corrects the generation of rotation gates like RX. Before this bugfix, there was a space between the name of the gate and the gate's angle. This problematic behavior would produce something like "rx (some_theta_value) q[some_target_qubit];" which is incorrect due to the space between rx and the opening parenthesis. The correct behavior introduced by this commit makes sure not to have that space there, so it produces something like "rx(some_theta_value) q[some_target_qubit];" which is valid QASM 2.0.